Moved templated expression code to expression.tpl.hpp. Minor cleanup of symbolic module.
