+++ source/symbolic/constant.hpp
@@ -48,6 +48,8 @@ template<class T> class Constant
     explicit Constant(const String& str, const T& value) : T(value), _name(str) { }
     const Identifier& name() const { return _name; }
     const T& value() const { return *this; }
+    friend OutputStream& operator<<(OutputStream& os, Constant<T> const& c) {
+        return os << c._name << "(=" << static_cast<T const&>(c) << ")"; }
   private:
     Identifier _name;
 };
